Add calculator functionality
